+++ rules/order/order.js
@@ -2,6 +2,10 @@ module.exports = [
   'custom-properties',
   'dollar-variables',
   'at-variables',
+  {
+    type: 'at-rule',
+    name: 'extend',
+  },
   'declarations',
   'rules',
   'at-rules'