+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Add `@extend` to order/order rule
 
 ## 0.2.0 - 2017-06-09
 - Upgrade dependencies